Try to get tote evaluator easing formula implemented. Working driving.
As of right now the tote easing formula does not work correct or at
least the change needs to be lowered since the interval in milliseconds
does not seem to be going through the function correctly.

The code for this probably should be implemented in a command instead of
being implemented using the JoystickButton.when*() commands since we do
not have the same amount of control as compared to when we simply has
the code interacting with the joysticks in the CollectTote command. I
figure we can at least create a joystick button with the correct mapping
in the OI class so that we can easily manage the joystick mappings.