Refactor project structure. Implement untested code for BinElevator.
The layout of where everything is located was somewhat confusing or
poorly named so a lot of things have been moved around to their
appropriately named class. Additionally, we are trying to use getters
instead of just `public static` so we don't need to worry about
accidentally setting to another instance or something stupid.

The ease-in function still does not seem to work from testing it out.