Use CustomDriveControl. More drive debugging
Since we need to be able to have more control over the driving control,
we have copied the stock default DriveControl and have stripped out
everything we are not using to make things simpler to deal with.

The preferences that get loaded onto SmartDashboard are using the
Preferences class from WPILib which will put the preferences in nice
grid table.

Add preferences to be able to disable components of the joystick drive
so that we can try and debug the drive control code.

Increase default tote lift speed from 0.5 to 0.75